requirements.txt


Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
gspread==5.10.0
Werkzeug==2.3.0

gunicorn==21.2.0

python-dotenv==1.0.0

google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
